backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic-settings==2.2.1
sqlalchemy==2.0.28
python-multipart==0.0.9
aiohttp==3.9.3
requests==2.31.0
python-dotenv==1.0.0

